bounty: NextJS deployment on Akash with SSL
Hi, 

I made an Akash manifest template and detailed walkthrough instructions for deploying a NextJS application on Akash: https://github.com/OIGbash/nextjs-on-akash

You can see the Akash deployment live here at its assigned URI: http://eg8thju72ldln5juaietlkqpeg.ingress.provider-0.prod.ams1.akash.pub/
And here at my domain with SSL enabled: https://nextjs-akash.com/
The site has a link to the directions on Github, but the directions can also be added there as a guide, and I'd be happy to keep the site live on Akash as an example if funded. I would also be happy to create a video walkthrough if that would be helpful. 

Have a great day!